store created by

+++ resources/views/users/index.blade.php
@@ -8,7 +8,9 @@
             <h2>Users Management</h2>
         </div>
         <div class="pull-right">
-            <a class="btn btn-success" href="{{ route('users.create') }}"> Create New User </a>
+        @if(auth()->user()->hasRole('Admin'))
+            <a class="btn btn-success" href="{{ route('users.create') }}">Create New User</a>
+        @endif
         </div>
     </div>
 </div>
@@ -42,11 +44,15 @@
       @endif
     </td>
     <td>
-       <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
-       <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
-        {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
-            {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-        {!! Form::close() !!}
+        @if(auth()->user()->hasRole('Admin'))
+            <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
+            <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+            {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+                {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+            {!! Form::close() !!}
+        @else
+            <span class="text-danger">You do not have permission to perform this action.</span>
+        @endif
     </td>
   </tr>
  @endforeach
